requirements.txt
uvicorn==0.15.0
fastapi==0.68.0
celery==5.1.2
redis==3.5.3
flower==1.0.0
psycopg2-binary==2.9.1
tortoise-orm==0.17.6
aerich==0.5.7
asyncpg==0.24.0